+++ pkgs/jnigen/lib/src/bindings/excluder.dart
@@ -7,8 +7,10 @@ import '../elements/elements.dart';
 import '../logging/logging.dart';
 import 'visitor.dart';
 
-bool _isPrivate(ClassMember classMember) =>
-    !classMember.isPublic && !classMember.isProtected;
+extension on ClassMember {
+  bool get isPrivate => !isPublic;
+  bool get hasPrivateName => name.startsWith('_');
+}
 
 class Excluder extends Visitor<Classes, void> {
   final Config config;
@@ -18,7 +20,8 @@ class Excluder extends Visitor<Classes, void> {
   @override
   void visit(Classes node) {
     node.decls.removeWhere((_, classDecl) {
-      final excluded = _isPrivate(classDecl) ||
+      final excluded = classDecl.isPrivate ||
+          classDecl.hasPrivateName ||
           !(config.exclude?.classes?.included(classDecl) ?? true);
       if (excluded) {
         log.fine('Excluded class ${classDecl.binaryName}');
@@ -40,23 +43,30 @@ class _ClassExcluder extends Visitor<ClassDecl, void> {
   @override
   void visit(ClassDecl node) {
     node.methods = node.methods.where((method) {
-      final included = !_isPrivate(method) &&
-          !method.name.startsWith('_') &&
-          method.name != '<clinit>' &&
-          (config.exclude?.methods?.included(node, method) ?? true);
-      if (!included) {
+      final isPrivate = method.isPrivate;
+      final hasPrivateName = method.hasPrivateName;
+      final isAbstractCtor = method.isCtor && node.isAbstract;
+      final isBridgeMethod = method.isSynthetic && method.isBridge;
+      final isExcludedInConfig =
+          config.exclude?.methods?.included(node, method) ?? false;
+      final excluded = isPrivate ||
+          hasPrivateName ||
+          isAbstractCtor ||
+          isBridgeMethod ||
+          isExcludedInConfig;
+      if (excluded) {
         log.fine('Excluded method ${node.binaryName}#${method.name}');
       }
-      return included;
+      return !excluded;
     }).toList();
     node.fields = node.fields.where((field) {
-      final included = !_isPrivate(field) &&
-          !field.name.startsWith('_') &&
-          (config.exclude?.fields?.included(node, field) ?? true);
-      if (!included) {
+      final excluded = field.isPrivate ||
+          field.hasPrivateName &&
+              (config.exclude?.fields?.included(node, field) ?? true);
+      if (excluded) {
         log.fine('Excluded field ${node.binaryName}#${field.name}');
       }
-      return included;
+      return !excluded;
     }).toList();
   }
 }