+++ pkgs/jnigen/lib/src/elements/elements.dart
@@ -164,7 +164,7 @@ class ClassDecl extends ClassMember implements Element<ClassDecl> {
   ClassDecl get classDecl => this;
 
   @override
-  String get name => finalName;
+  String get name => binaryName.split('.').last;
 
   bool get isObject => superCount == 0;
 
@@ -443,10 +443,13 @@ abstract class ClassMember {
   ClassDecl get classDecl;
   Set<String> get modifiers;
 
+  bool get isAbstract => modifiers.contains('abstract');
   bool get isStatic => modifiers.contains('static');
   bool get isFinal => modifiers.contains('final');
   bool get isPublic => modifiers.contains('public');
   bool get isProtected => modifiers.contains('protected');
+  bool get isSynthetic => modifiers.contains('synthetic');
+  bool get isBridge => modifiers.contains('bridge');
 }
 
 @JsonSerializable(createToJson: false)