+++ pkgs/jnigen/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.11.1-wip
+
+- Fixed an issue where inheriting a generic class could generate incorrect code.
+- No longer generating constructors for abstract classes.
+- No longer generating `protected` elements.
+
 ## 0.11.0
 
 - No changes. Keep major version in sync with `package:jni`.